requirements.txt
discord.py
bs4
